datapipelines/requirements.txt
pydrive==1.3.1
requests==2.31.0
pandas==2.0.3
pandas-gbq==0.19.2
openpyxl==3.1.2
python-slugify==8.0.1
